# Optimize performance when Git is used as storage repository