--- a/src/archivist/recorder/repositories/git/index.js
+++ b/src/archivist/recorder/repositories/git/index.js
@@ -66,8 +66,13 @@ export default class GitRepository extends RepositoryInterface {
   }
 
   async findLatest(serviceId, termsType, documentId) {
-    const filePath = DataMapper.generateFilePath(serviceId, termsType, documentId);
-    const commit = await this.git.getCommit([filePath]);
+    const matchingFilesPaths = await this.git.listFiles(DataMapper.generateFilePath(serviceId, termsType, documentId));
+
+    if (!matchingFilesPaths.length) {
+      return null;
+    }
+
+    const commit = await this.git.getCommit([...matchingFilesPaths]); // Returns the most recent commit that modified any of the matching files. If multiple files match the path pattern (e.g. both HTML and PDF versions exist), returns the commit that last modified any of them
 
     return this.#toDomain(commit);
   }