This PR implements several optimizations to improve performance and memory usage:

### Technical improvements:

- Replace wildcard patterns in `git log` commands with specific paths
- Enable `commit-graph` for faster commit history traversal (fixes #1101)
- Release HTML/PDF content from memory after term extraction to reduce memory footprint
- Remove unnecessary commit hooks

### Performance benchmarks:

Testing was performed on two distinct datasets:

#### ToS;DR Collection
- Dataset: 100 randomly selected services
- Source: [ToS;DR declarations](https://github.com/tosdr/tosdr-declarations) with [full history](https://github.com/tosdr/tosdr-snapshots)
- Scale: ~92,000 commits, 18,083 terms from 9,202 services

| Operation | Before | After | Improvement |
|-----------|--------|-------|-------------|
| Extraction | 3h 17m 42s | 42s | 99.6% faster |
| Tracking | 1h 19m 8s | 4m 52s | 93.8% faster |

#### P2B Compliance Collection
- Dataset: Complete collection
- Source: [P2B-compliance declarations](https://github.com/OpenTermsArchive/p2b-compliance-declarations) with [full history](https://github.com/OpenTermsArchive/p2b-compliance-snapshots)
- Scale: ~385,000 commits, 245 terms from 203 services

| Operation | Before | After | Improvement |
|-----------|--------|-------|-------------|
| Extraction | 10m 50s | 1m 25s | 87.0% faster |
| Tracking | 16m 15s | 11m 15s | 30.8% faster |

The optimizations show greater impact on collections with larger numbers of declared services and terms.

Fixes #1101. Special thanks to @michielbdejong for reporting the issue, investigating, and providing valuable suggestions 🙏